chore: cleanup on playlist table
